lil user manual so user can see available commands
also made this so we can easily see the available commands, more so to see what we've done and what needs doing.

will probs make more detailed manual pages for diff commands and change this manual page to just show a list of commands, and if you want more detailed list type "man <commandname>" like the one in powershell, but right now this is fine lol.